feat: Refactor Overview component into modular architecture with reusable components, custom hooks, optimized performance, and improved maintainability
