+++ frontend/src/components/Pages/Dashboard/Overview/index.jsx
@@ -1,0 +1,19 @@
+// Main Overview component export
+export { default } from './Overview'
+
+// Named exports for individual components (optional - for flexibility)
+export { default as DashboardHeader } from './components/DashboardHeader'
+export { default as DashboardCards } from './components/DashboardCards'
+export { default as GameWorlds } from './components/GameWorlds'
+export { default as LoadingSpinner } from './components/LoadingSpinner'
+export { ConfirmationModal, LeaderboardModal } from './components/Modals'
+
+// Custom hooks exports
+export { useOverviewData } from './hooks/useOverviewData'
+export { useWorldStatus } from './hooks/useWorldStatus'
+
+// Utility functions export
+export * from './utils'
+
+// Constants export
+export * from './constants'
